+++ pkg/storegateway/parquet_bucket_store.go
@@ -25,6 +25,7 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
+	"github.com/cortexproject/cortex/pkg/util/spanlogger"
 	"github.com/cortexproject/cortex/pkg/util/validation"
 )
 
@@ -73,12 +74,14 @@ func (p *parquetBucketStore) findParquetBlocks(ctx context.Context, blockMatcher
 
 // Series implements the store interface for a single parquet bucket store
 func (p *parquetBucketStore) Series(req *storepb.SeriesRequest, srv storepb.Store_SeriesServer) (err error) {
+	spanLog, ctx := spanlogger.New(srv.Context(), "ParquetBucketStore.Series")
+	defer spanLog.Finish()
+
 	matchers, err := storecache.MatchersToPromMatchersCached(p.matcherCache, req.Matchers...)
 	if err != nil {
 		return status.Error(codes.InvalidArgument, err.Error())
 	}
 
-	ctx := srv.Context()
 	resHints := &hintspb.SeriesResponseHints{}
 	var anyHints *types.Any
 
@@ -158,6 +161,9 @@ func (p *parquetBucketStore) Series(req *storepb.SeriesRequest, srv storepb.Stor
 
 // LabelNames implements the store interface for a single parquet bucket store
 func (p *parquetBucketStore) LabelNames(ctx context.Context, req *storepb.LabelNamesRequest) (*storepb.LabelNamesResponse, error) {
+	spanLog, ctx := spanlogger.New(ctx, "ParquetBucketStore.LabelNames")
+	defer spanLog.Finish()
+
 	matchers, err := storecache.MatchersToPromMatchersCached(p.matcherCache, req.Matchers...)
 	if err != nil {
 		return nil, status.Error(codes.InvalidArgument, err.Error())
@@ -213,6 +219,9 @@ func (p *parquetBucketStore) LabelNames(ctx context.Context, req *storepb.LabelN
 
 // LabelValues implements the store interface for a single parquet bucket store
 func (p *parquetBucketStore) LabelValues(ctx context.Context, req *storepb.LabelValuesRequest) (*storepb.LabelValuesResponse, error) {
+	spanLog, ctx := spanlogger.New(ctx, "ParquetBucketStore.LabelValues")
+	defer spanLog.Finish()
+
 	matchers, err := storecache.MatchersToPromMatchersCached(p.matcherCache, req.Matchers...)
 	if err != nil {
 		return nil, status.Error(codes.InvalidArgument, err.Error())